requirements.txt
certifi==2022.12.7
click==8.1.3
distlib==0.3.6
filelock==3.8.2
Flask==2.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.21.6
pandas==1.3.5
pipenv==2022.12.19
platformdirs==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
SQLAlchemy==1.4.45
virtualenv==20.17.1
virtualenv-clone==0.5.7
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
